develpment commit, not runable yet

+++ source/performancetests/main.cpp
@@ -14,7 +14,7 @@ int main()
 	crap::binary_tree<i32>* bin_tree = new crap::binary_tree<i32>();
 
     std::cout << "\t inserting 1000 integers" << std::endl;
-	time = crap::time::current_tick();
+	time = (i32)crap::time::current_tick();
     for(int i=0; i<1000; ++i)
 	{
 		bin_tree->insert(i);
@@ -22,12 +22,12 @@ int main()
 	std::cout << "\t\t"<< (f32)( (i32)crap::time::current_tick() - time ) / crap::time::TICKS_PER_SECOND << " seconds" << std::endl;
 
 	std::cout << "\t realigning tree" << std::endl;
-	time = crap::time::current_tick();
+	time = (i32)crap::time::current_tick();
 	bin_tree->realign();
 	std::cout << "\t\t"<< (f32)( (i32)crap::time::current_tick() - time ) / crap::time::TICKS_PER_SECOND << " seconds" << std::endl;
 
     std::cout << "\t searching 1000 random integers" << std::endl;
-	time = crap::time::current_tick();
+	time = (i32)crap::time::current_tick();
     for(int i=0; i<1000; ++i)
 	{
         bin_tree->find( crap::random<int>::range(0, 999) );
@@ -35,7 +35,7 @@ int main()
 	std::cout << "\t\t"<< (f32)( (i32)crap::time::current_tick() - time ) / crap::time::TICKS_PER_SECOND << " seconds" << std::endl;
 
     std::cout << "\t deleting 1000 integers" << std::endl;
-	time = crap::time::current_tick();
+	time = (i32)crap::time::current_tick();
     for(int i=0; i<1000; ++i)
 	{
 		bin_tree->remove(i);
